move models/searchPageState to state/searchPageState
