+++ assets/src/components/searchPage.tsx
@@ -3,9 +3,9 @@ import React, { ReactElement, useContext, useState } from "react"
 import { SocketContext } from "../contexts/socketContext"
 import { StateDispatchContext } from "../contexts/stateDispatchContext"
 import useSearchResults from "../hooks/useSearchResults"
-import { SearchPageState } from "../models/searchPageState"
 import { isVehicle } from "../models/vehicle"
 import { Vehicle, VehicleId, VehicleOrGhost } from "../realtime"
+import { SearchPageState } from "../state/searchPageState"
 import Map from "./map"
 import PropertiesPanel from "./propertiesPanel"
 import RecentSearches from "./recentSearches"
