+++ assets/src/components/searchResults.tsx
@@ -1,9 +1,9 @@
 import React, { useContext } from "react"
 import { StateDispatchContext } from "../contexts/stateDispatchContext"
-import { setSearchText } from "../models/searchPageState"
 import { isVehicle } from "../models/vehicle"
 import { Vehicle, VehicleOrGhost } from "../realtime"
 import { selectVehicle } from "../state"
+import { setSearchText } from "../state/searchPageState"
 import PropertiesList from "./propertiesList"
 import { RouteVariantName } from "./routeVariantName"
